+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.5.1 / _Unreleased_
 
+- Fix Chef provisioner configuration if a proxy is set to `false` ([GH-24][])
 
 # 0.5.0 / 2013-09-11
 
@@ -54,3 +55,4 @@
 [GH-17]: https://github.com/tmatilai/vagrant-proxyconf/issues/17 "Issue 17"
 [GH-19]: https://github.com/tmatilai/vagrant-proxyconf/issues/19 "Issue 19"
 [GH-21]: https://github.com/tmatilai/vagrant-proxyconf/issues/21 "Issue 21"
+[GH-24]: https://github.com/tmatilai/vagrant-proxyconf/issues/24 "Issue 24"